Added post type management
Plugin settings set to top-level in admin menu, and post type management moved to submenu beneath that.  I removed the "enable post type" functionality because it felt excessive along with the ability to create a custom type.  The icon table has been reformatted to inform users of the icons that will be displayed in the resource center according to post slug.